refactor: restructure project to align with domain context
