+++ next.config.ts
@@ -1,7 +1,7 @@
 import type { NextConfig } from 'next'
 import createNextIntlPlugin from 'next-intl/plugin'
 
-const withNextIntl = createNextIntlPlugin()
+const withNextIntl = createNextIntlPlugin('./src/app/(i18n)/request.ts')
 
 const nextConfig: NextConfig = {
   /* Next.js configurations */
