+++ src/app/(i18n)/request.ts
@@ -9,6 +9,6 @@ export default getRequestConfig(async ({ requestLocale }) => {
 
   return {
     locale,
-    messages: (await import(`../../public/locales/${locale}.json`)).default,
+    messages: (await import(`../../../public/locales/${locale}.json`)).default,
   }
 })
