+++ src/app/(routes)/[lang]/(landing)/layout.tsx
