+++ src/app/(routes)/[lang]/(landing)/page.tsx
