+++ src/contexts/listings/infra/InMemoryListingRepository.ts
@@ -1,9 +1,9 @@
 import { randomUUID } from 'node:crypto'
 import { Listing, ListingConsoleHealth } from '@/listings/domain/Listing'
 import { ListingRepository } from '@/listings/domain/ListingRepository'
-import { Random } from '@/shared/utils/Random'
 import { faker } from '@faker-js/faker/locale/en'
 import { Money } from 'ts-money'
+import { Random } from '../../shared/domain/Random'
 
 export class InMemoryListingRepository implements ListingRepository {
   private static instance: InMemoryListingRepository
