+++ src/contexts/shared/components/organisms/NavBar/NavBar.tsx
@@ -5,7 +5,7 @@ export const NavBar = () => {
   return (
     <Flex className={styles['navBar']} asChild p="4">
       <nav>
-        <Link href="/" underline="none" weight="bold" size="5">
+        <Link href="/public" underline="none" weight="bold" size="5">
           Enriconsolas
         </Link>
       </nav>
